+++ NEWS.md
@@ -3,6 +3,7 @@
 ## ggauto 0.0.1.9000
 
 * Add CRAN badge
+* Fix: character ordering
 
 ## ggauto 0.0.1